// .temp/packages/types/generated.ts
export type User = {
    id: string;
    avatar_url: string;
    avatar_urls: AvatarUrls;
    cover_photo_url: string | null;
    email: string;
    username: string;
    display_name: string;
    system: boolean;
    integration: boolean;
    notifications_paused: boolean;
    notification_pause_expires_at: string | null;
    timezone: string | null;
    logged_in: boolean;
    type_name: string;
};
export type OrganizationMembershipStatus = {
    message: string;
    emoji: string;
    expiration_setting: '30m' | '1h' | '4h' | 'today' | 'this_week' | 'custom';
    expires_at: string | null;
    pause_notifications: boolean;
    expires_in: '30m' | '1h' | '4h' | 'today' | 'this_week' | 'custom';
};
export type OrganizationMember = {
    id: string;
    role: 'admin' | 'member' | 'viewer' | 'guest';
    created_at: string;
    deactivated: boolean;
    is_organization_member: boolean;
    user: User;
    status: OrganizationMembershipStatus | null;
};
export type ImageUrls = {
    original_url: string;
    thumbnail_url: string;
    feed_url: string;
    email_url: string;
    slack_url: string;
    large_url: string;
};
export type Attachment = {
    id: string;
    file_type: string;
    url: string;
    app_url: string;
    download_url: string;
    preview_url: string | null;
    preview_thumbnail_url: string | null;
    image_urls: ImageUrls | null;
    link: boolean;
    image: boolean;
    video: boolean;
    audio: boolean;
    origami: boolean;
    principle: boolean;
    lottie: boolean;
    stitch: boolean;
    gif: boolean;
    duration: number;
    width: number;
    height: number;
    subject_type: string | null;
    name: string | null;
    size: number | null;
    remote_figma_url: string | null;
    no_video_track: boolean;
    gallery_id: string | null;
    type_name: string;
    subject_id: string | null;
    is_subject_comment: boolean;
    relative_url: string;
    preview_relative_url: string | null;
    comments_count: number;
    key: string | null;
    optimistic_id?: string | null;
    optimistic_file_path?: string | null;
    optimistic_preview_file_path?: string | null;
    optimistic_imgix_video_file_path?: string | null;
    optimistic_src?: string | null;
    optimistic_preview_src?: string | null;
    optimistic_ready: boolean;
    client_error?: string | null;
};
// ...
export type CallPeer = {
    member: OrganizationMember;
    active: boolean;
    remote_peer_id: string;
};
// ...
export type SlackChannel = {
    id: string;
    name: string;
    is_private: boolean;
};
export type ProjectDisplayPreference = {
    display_reactions: boolean;
    display_attachments: boolean;
    display_comments: boolean;
    display_resolved: boolean;
};
export type Project = {
    id: string;
    name: string;
    description: string | null;
    created_at: string;
    archived_at: string | null;
    archived: boolean;
    last_activity_at: string;
    slack_channel_id: string | null;
    posts_count: number;
    cover_photo_url: string | null;
    url: string;
    accessory: string | null;
    private: boolean;
    personal: boolean;
    is_general: boolean;
    is_default: boolean;
    contributors_count: number;
    members_and_guests_count: number;
    members_count: number;
    guests_count: number;
    call_room_url: string | null;
    message_thread_id: string | null;
    organization_id: string;
    viewer_has_favorited: boolean;
    viewer_can_archive: boolean;
    viewer_can_destroy: boolean;
    viewer_can_unarchive: boolean;
    viewer_can_update: boolean;
    viewer_has_subscribed: boolean;
    viewer_subscription: 'posts_and_comments' | 'new_posts' | 'none';
    viewer_is_member: boolean;
    unread_for_viewer: boolean;
    slack_channel: SlackChannel | null;
    type_name: string;
    viewer_display_preferences: ProjectDisplayPreference | null;
    display_preferences: ProjectDisplayPreference;
};
// ...
export type MiniProject = {
    id: string;
    name: string;
    accessory: string | null;
    private: boolean;
    archived: boolean;
    message_thread_id: string | null;
};
export type NotificationTarget = {
    id: string;
    type: string;
    title: string;
    project: MiniProject | null;
    resolved: boolean;
};
export type SummaryBlock = {
    text?: {
        content: string;
        bold?: boolean;
        nowrap?: boolean;
    };
    img?: {
        src: string;
        alt: string;
    };
};
export type FollowUp = {
    id: string;
    show_at: string;
    inbox_key: string;
    organization_slug: string;
    member: OrganizationMember;
    subject: FollowUpSubject;
    target: NotificationTarget;
    summary_blocks: SummaryBlock[];
    belongs_to_viewer: boolean;
    type_name: string;
};
export type OrganizationCallFollowUpPostRequest = {
    show_at: string;
};
export type PublicOrganization = {
    id: string;
    avatar_url: string;
    avatar_urls: AvatarUrls;
    name: string;
    slug: string;
    viewer_is_admin: boolean;
    viewer_can_leave: boolean;
};
export type PostLink = {
    id: string;
    name: string;
    url: string;
};
export type Tag = {
    id: string;
    name: string;
    posts_count: number;
    url: string;
    viewer_can_destroy: boolean;
};
export type PollOption = {
    id: string;
    description: string;
    votes_count: number;
    votes_percent: number;
    viewer_voted: boolean;
};
export type Poll = {
    id: string;
    description: string;
    votes_count: number;
    options: PollOption[];
    viewer_voted: boolean;
};
export type FeedbackRequest = {
    id: string;
    has_replied: boolean;
    member: OrganizationMember;
};
export type SubjectFollowUp = {
    id: string;
    member: OrganizationMember;
    show_at: string;
    belongs_to_viewer: boolean;
};
export type ResolvedComment = {
    id: string;
    created_at: string;
    body_html: string;
    url: string;
    viewer_is_author: boolean;
    member: OrganizationMember;
};
export type SyncCustomReaction = {
    id: string;
    name: string;
    file_url: string;
    created_at: string;
};
export type GroupedReaction = {
    viewer_reaction_id: string | null;
    emoji: string | null;
    tooltip: string;
    reactions_count: number;
    custom_content: SyncCustomReaction | null;
};
export type Commenters = {
    latest_commenters: OrganizationMember[];
};
export type PostResolution = {
    resolved_at: string;
    resolved_by: OrganizationMember;
    resolved_html: string | null;
    resolved_comment: ResolvedComment | null;
};
export type ResourceMentionPost = {
    id: string;
    title: string;
    created_at: string;
    published_at: string | null;
    url: string;
};
export type ResourceMentionCall = {
    id: string;
    title: string;
    created_at: string;
    url: string;
};
export type ResourceMentionNote = {
    id: string;
    title: string;
    created_at: string;
    url: string;
};
export type ResourceMention = {
    id: string;
    post: ResourceMentionPost | null;
    call: ResourceMentionCall | null;
    note: ResourceMentionNote | null;
    type_name: string;
};
export type Post = {
    id: string;
    title: string;
    is_title_from_description: boolean;
    created_at: string;
    published_at: string | null;
    published: boolean;
    last_activity_at: string;
    comments_count: number;
    resolved_comments_count: number;
    version: number;
    path: string;
    channel_name: string;
    views_count: number;
    non_member_views_count: number;
    status: 'none' | 'feedback_requested';
    visibility: 'default' | 'public';
    open_graph_image_url: string | null;
    thumbnail_url: string | null;
    unfurled_link: string | null;
    description_html: string;
    truncated_description_html: string;
    is_text_content_truncated: boolean;
    truncated_description_text: string;
    url: string;
    type_name: string;
    organization: PublicOrganization;
    attachments: Attachment[];
    links: PostLink[];
    tags: Tag[];
    poll: Poll | null;
    feedback_requests: FeedbackRequest[] | null;
    follow_ups: SubjectFollowUp[];
    member: OrganizationMember;
    resolved_comment: ResolvedComment | null;
    grouped_reactions: GroupedReaction[];
    project: MiniProject;
    has_parent: boolean;
    has_iterations: boolean;
    viewer_is_organization_member: boolean;
    viewer_is_author: boolean;
    viewer_has_commented: boolean;
    preview_commenters: Commenters;
    viewer_feedback_status: 'none' | 'viewer_requested' | 'open';
    viewer_has_subscribed: boolean;
    viewer_has_viewed: boolean;
    viewer_has_favorited: boolean;
    unseen_comments_count: number;
    viewer_can_resolve: boolean;
    viewer_can_favorite: boolean;
    viewer_can_edit: boolean;
    viewer_can_delete: boolean;
    viewer_can_create_issue: boolean;
    resolution: PostResolution | null;
    latest_comment_preview: string | null;
    latest_comment_path: string | null;
    viewer_is_latest_comment_author: boolean;
    project_pin_id: string | null;
    resource_mentions: ResourceMention[];
};
export type Permission = {
    id: string;
    user: User;
    action: 'view' | 'edit';
};
export type Note = {
    id: string;
    title: string;
    created_at: string;
    last_activity_at: string;
    content_updated_at: string;
    comments_count: number;
    resolved_comments_count: number;
    channel_name: string;
    presence_channel_name: string;
    description_thumbnail_base_url: string | null;
    public_visibility: boolean;
    non_member_views_count: number;
    description_html: string;
    description_state: string | null;
    project: Project | null;
    follow_ups: FollowUp[];
    type_name: string;
    url: string;
    public_share_url: string;
    project_permission: 'none' | 'view' | 'edit';
    member: OrganizationMember;
    viewer_is_author: boolean;
    viewer_can_comment: boolean;
    viewer_can_edit: boolean;
    viewer_can_delete: boolean;
    viewer_has_favorited: boolean;
    latest_commenters: OrganizationMember[];
    permitted_users: Permission[];
    project_pin_id: string | null;
    resource_mentions: ResourceMention[];
};
export type Call = {
    id: string;
    title: string | null;
    summary_html: string | null;
    is_edited: boolean;
    created_at: string;
    started_at: string;
    stopped_at: string | null;
    duration: string | null;
    recordings_duration: string | null;
    active: boolean;
    project_permission: 'none' | 'view' | 'edit';
    channel_name: string;
    peers: CallPeer[];
    project: MiniProject | null;
    follow_ups: SubjectFollowUp[];
    type_name: string;
    viewer_can_edit: boolean;
    viewer_can_destroy_all_recordings: boolean;
    viewer_has_favorited: boolean;
    processing_generated_title: boolean;
    processing_generated_summary: boolean;
    project_pin_id: string | null;
    url: string;
};
// ...
export type ExternalRecord = {
    created_at: string;
    remote_record_id: string;
    remote_record_title: string;
    remote_record_url: string;
    service: string;
    type: string;
    linear_issue_identifier: string;
    linear_issue_state: LinearIssueState;
    linear_identifier: string;
    linear_state: LinearIssueState;
};
// ...
export type TimelineEvent = {
    id: string;
    created_at: string;
    action: 'post_resolved' | 'post_unresolved' | 'post_visibility_updated' | 'post_referenced_in_external_record' | 'created_linear_issue_from_post' | 'comment_referenced_in_external_record' | 'created_linear_issue_from_comment' | 'subject_project_updated' | 'subject_referenced_in_internal_record' | 'subject_pinned' | 'subject_unpinned' | 'subject_title_updated';
    subject_updated_from_title: string | null;
    subject_updated_to_title: string | null;
    subject_updated_from_project: MiniProject | null;
    subject_updated_to_project: MiniProject | null;
    comment_reference_subject_type: string | null;
    comment_reference_subject_title: string | null;
    member_actor: OrganizationMember | null;
    external_reference: ExternalRecord | null;
    post_reference: Post | null;
    comment_reference: Comment | null;
    note_reference: Note | null;
};
export type Comment = {
    id: string;
    created_at: string;
    timestamp: number | null;
    x: number | null;
    y: number | null;
    body_html: string;
    note_highlight: string | null;
    resolved_at: string | null;
    resolved_by: OrganizationMember | null;
    type_name: string;
    subject_type: string;
    subject_id: string;
    url: string;
    viewer_can_resolve: boolean;
    viewer_can_create_issue: boolean;
    attachment_id: string | null;
    canvas_preview_url: string | null;
    attachment_thumbnail_url: string | null;
    viewer_is_author: boolean;
    viewer_can_edit: boolean;
    viewer_can_follow_up: boolean;
    viewer_can_react: boolean;
    viewer_can_delete: boolean;
    member: OrganizationMember;
    attachments: Attachment[];
    grouped_reactions: GroupedReaction[];
    replies: Comment[];
    follow_ups: SubjectFollowUp[];
    parent_id: string | null;
    is_optimistic: boolean;
    optimistic_id: string | null;
    timeline_events: TimelineEvent[];
    resource_mentions: ResourceMention[];
};